Refactor:Refactor element APIs to api folder
