+++ src/newCategoryForm.js
@@ -1,8 +1,9 @@
-import { form, textInput, submitButton } from './formTags.js';
+import { form, textInput, submitButton } from "./api/formTags.js";
 
-const newCategoryForm = () => form([
-  textInput('Enter category name', '', { name: 'name' }),
-  submitButton(),
-]);
+const newCategoryForm = () =>
+  form([
+    textInput("Enter category name", "", { name: "name" }),
+    submitButton(),
+  ]);
 
 export default newCategoryForm;
